chore: remove unused props
