+++ src/app/App.tsx
@@ -9,6 +9,16 @@ import {
 import '../store';
 import '../styles/app.scss';
 
+/**
+ * App Component.
+ *
+ * This function returns a JSX component that comprises
+ * the Import Button and Fields.
+ *
+ * @since 1.0.0
+ *
+ * @return {JSX.Element} The App component.
+ */
 const App = (): JSX.Element => {
 	return (
 		<main>
