+++ src/components/ImportButton.tsx
@@ -1,21 +1,8 @@
 import { __ } from '@wordpress/i18n';
 import { Button } from '@wordpress/components';
-import type { MediaFrame } from '@wordpress/media-utils';
-import { useSelect, useDispatch } from '@wordpress/data';
-import apiFetch from '@wordpress/api-fetch';
+import { useSelect } from '@wordpress/data';
 
-import { getModalParams } from '../utils';
-
-interface ImportButtonProps {
-	parsedSQL: {
-		tableName: string;
-		tableRows: any[];
-		tableColumns: string[];
-	};
-	setIsLoading: Function;
-	setSqlNotice: Function;
-	setParsedSQL: Function;
-}
+import { handleImport, handleUpload } from '../utils';
 
 /**
  * Import Button Component.
@@ -28,8 +15,6 @@ interface ImportButtonProps {
  * @return {JSX.Element} The Import Button component.
  */
 const ImportButton = (): JSX.Element => {
-	const { setIsLoading, setSqlNotice, setParsedSQL } =
-		useDispatch( 'sql-to-cpt' );
 	const { parsedSQL } = useSelect( ( select ) => {
 		const store: any = select( 'sql-to-cpt' );
 
@@ -38,104 +23,22 @@ const ImportButton = (): JSX.Element => {
 		};
 	}, [] );
 
-	/**
-	 * Handle Upload.
-	 *
-	 * This function is responsible for opening the
-	 * WP media modal to enable user select.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @return {void}
-	 */
-	const handleUpload = (): void => {
-		const wpMediaModal = wp.media( getModalParams() );
-		wpMediaModal.on( 'select', () => handleSelect( wpMediaModal ) ).open();
-	};
-
-	/**
-	 * Handle Selection.
-	 *
-	 * This function is responsible for handling a
-	 * selection made by the user.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param {MediaFrame} wpMediaModal WP Media Modal.
-	 * @return Promise<void>
-	 */
-	const handleSelect = async (
-		wpMediaModal: MediaFrame
-	): Promise< void > => {
-		const args = wpMediaModal.state().get( 'selection' ).first().toJSON();
-
-		// Reset.
-		setSqlNotice( '' );
-		setParsedSQL( {
-			tableName: '',
-			tableColumns: [],
-			tableRows: [],
-		} );
-		setIsLoading( true );
-
-		// Parse SQL.
-		try {
-			setParsedSQL(
-				await apiFetch( {
-					path: '/sql-to-cpt/v1/parse',
-					method: 'POST',
-					data: {
-						...args,
-					},
-				} )
-			);
-			setIsLoading( false );
-		} catch ( { message } ) {
-			setIsLoading( false );
-			setSqlNotice( message );
-		}
-	};
-
-	/**
-	 * Handle Import.
-	 *
-	 * This function is responsible for handling the
-	 * import made by the user.
-	 *
-	 * @since 1.1.0
-	 *
-	 * @return Promise<void>
-	 */
-	const handleImport = async (): Promise< void > => {
-		setIsLoading( true );
-
-		// Import SQL.
-		try {
-			const url = await apiFetch( {
-				path: '/sql-to-cpt/v1/import',
-				method: 'POST',
-				data: {
-					...parsedSQL,
-				},
-			} );
-			if ( url ) {
-				window.location.href = `${ url }`;
-			}
-			setIsLoading( false );
-		} catch ( { message } ) {
-			setIsLoading( false );
-			setSqlNotice( message );
-		}
-	};
-
 	return (
 		<>
-			{ parsedSQL.tableRows.length < 1 ? (
-				<Button variant="primary" onClick={ handleUpload }>
+			{ ( parsedSQL?.tableRows?.length || 0 ) < 1 ? (
+				<Button
+					data-testid="upload"
+					variant="primary"
+					onClick={ handleUpload }
+				>
 					{ __( 'Upload SQL File', 'sql-to-cpt' ) }
 				</Button>
 			) : (
-				<Button variant="primary" onClick={ handleImport }>
+				<Button
+					data-testid="convert"
+					variant="primary"
+					onClick={ handleImport }
+				>
 					{ __( 'Convert to CPT', 'sql-to-cpt' ) }
 				</Button>
 			) }
